requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.3.0
pydantic==2.8.2
pydantic_core==2.20.1
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.32.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
yookassa==3.3.0
